Add files via upload
This project is called Interpreter project. 
The purpose of this project is to be able to implement an interpreter for the mock language X.
The interpreter attempt to execute the bytecode file that contains sources of bytecodes
(byte-code interpreter). That bytecodes execute to run the bytecode loader(initialization)
and create the program in the virtual machine. That virtual machine leads the program to
run the interpreter bytecodes from the corresponding bytecode loaders.